+++ eslint.config.js
@@ -6,7 +6,7 @@ export default defineConfig(
   eslint.configs.recommended,
   tseslint.configs.recommended,
   {
-    ignores: ['*.*', '**/*.js', 'dist', 'fixtures', 'scripts'],
+    ignores: ['*.js', '*.ts', '**/*.js', 'dist', 'fixtures', 'scripts'],
   },
   {
     files: ['src/**/*.ts'],