+++ src/test/unit/tsCodeAction/transform.test.ts
@@ -17,7 +17,7 @@ import {
   ternaryToSwitch,
   ternaryToSwitchLeft,
   whileToDoWhile,
-} from '@/tsCodeAction/transform'
+} from '@/codeAction/javascript/transform'
 import assert from 'assert/strict'
 import { fileURLToPath } from 'url'
 import { Language, Parser } from 'web-tree-sitter'