+++ src/util/logger.ts
@@ -1,0 +1,5 @@
+import { window } from 'vscode'
+
+export const logger = window.createOutputChannel('Make VSCode Extension', {
+  log: true,
+})